+++ WindowsAppProject/course_dashboard.resx
